moved UART_handler class out of hardware.py into a its own device file
